Pull up context menu to tabbed data grid
