+++ web/src/features/moreCast2/components/ForecastSummaryDataGrid.tsx
@@ -5,16 +5,21 @@ import { ModelChoice, ModelType } from 'api/moreCast2API'
 import { MoreCast2Row } from 'features/moreCast2/interfaces'
 import { LinearProgress, Menu, MenuItem } from '@mui/material'
 import ApplyToColumnMenu from 'features/moreCast2/components/ApplyToColumnMenu'
-import { isEqual } from 'lodash'
 import { DataGridColumns } from 'features/moreCast2/components/DataGridColumns'
 
 interface ForecastSummaryDataGridProps {
   loading: boolean
   rows: MoreCast2Row[]
   clickedColDef: GridColDef | null
+  contextMenu: {
+    mouseX: number
+    mouseY: number
+  } | null
   onCellEditStop: (value: boolean) => void
   setClickedColDef: React.Dispatch<React.SetStateAction<GridColDef | null>>
   updateColumnWithModel: (modelType: ModelType, colDef: GridColDef) => void
+  handleColumnHeaderClick: GridEventListener<'columnHeaderClick'>
+  handleClose: () => void
 }
 
 const useStyles = makeStyles({
@@ -28,27 +33,13 @@ const ForecastSummaryDataGrid = ({
   loading,
   rows,
   clickedColDef,
+  contextMenu,
   onCellEditStop,
-  setClickedColDef,
-  updateColumnWithModel
+  updateColumnWithModel,
+  handleColumnHeaderClick,
+  handleClose
 }: ForecastSummaryDataGridProps) => {
   const classes = useStyles()
-
-  const [contextMenu, setContextMenu] = React.useState<{
-    mouseX: number
-    mouseY: number
-  } | null>(null)
-
-  const handleColumnHeaderClick: GridEventListener<'columnHeaderClick'> = (params, event) => {
-    if (!isEqual(params.colDef.field, 'stationName') && !isEqual(params.colDef.field, 'forDate')) {
-      setClickedColDef(params.colDef)
-      setContextMenu(contextMenu === null ? { mouseX: event.clientX, mouseY: event.clientY } : null)
-    }
-  }
-
-  const handleClose = () => {
-    setContextMenu(null)
-  }
 
   return (
     <div className={classes.root} data-testid={`morecast2-data-grid`}>
